Moves directory util functions to their own class
